requirements.txt
numpy>=1.9.0
scipy>=0.14.0
scikit-learn>=0.15.2
Pillow>=2.6.0